+++ packages/syft/src/syft/__init__.py
@@ -29,7 +29,6 @@ To begin your education in Syft, continue to the :py:mod:`syft.core.node.vm.vm` 
 __version__ = "0.8.0-beta.0"
 
 # stdlib
-import os
 from pathlib import Path
 import sys
 from typing import Any
@@ -62,6 +61,7 @@ from .core.node.new.user_code import UserCodeStatus  # noqa: F401
 from .core.node.new.user_code import syft_function  # noqa: F401
 from .core.node.new.user_service import UserService  # noqa: F401
 from .core.node.worker import Worker  # noqa: F401
+from .external import enable_external_lib  # noqa: F401
 from .registry import DomainRegistry  # noqa: F401
 from .registry import NetworkRegistry  # noqa: F401
 from .search import Search  # noqa: F401
@@ -77,17 +77,6 @@ requires = make_requires(LATEST_STABLE_SYFT, __version__)
 sys.path.append(str(Path(__file__)))
 
 logger.start()
-
-OBLV = os.getenv("INSTALL_OBLV_CLI", "false") == "true"
-if OBLV:
-    # relative
-    from .oblv import OblvEnclavePointer  # noqa: F401
-    from .oblv import check_oblv_proxy_installation_status  # noqa: F401
-    from .oblv import create_deployment  # noqa: F401
-    from .oblv import create_oblv_key_pair  # noqa: F401
-    from .oblv import get_oblv_public_key  # noqa: F401
-    from .oblv import install_oblv_proxy  # noqa: F401
-    from .oblv import login as oblv_login  # noqa: F401
 
 
 def module_property(func: Any) -> None:
