+++ packages/syft/src/syft/external/oblv/__init__.py
@@ -1,7 +1,6 @@
 # relative
 from .auth import login  # noqa: F401
 from .deployment import create_deployment  # noqa: F401
-from .oblv_enclave_pointer import OblvEnclavePointer  # noqa: F401
 from .oblv_proxy import check_oblv_proxy_installation_status  # noqa: F401
 from .oblv_proxy import create_oblv_key_pair  # noqa: F401
 from .oblv_proxy import get_oblv_public_key  # noqa: F401
