+++ packages/syft/src/syft/external/oblv/deployment.py
@@ -7,7 +7,7 @@ from oblv.models import CreateDeploymentInput
 import yaml
 
 # relative
-from ..util import bcolors
+from ...util import bcolors
 from .auth import login
 from .constants import INFRA
 from .constants import REF
