+++ packages/syft/src/syft/external/oblv/oblv_keys.py
@@ -1,7 +1,7 @@
 # relative
-from ..serializable import serializable
-from ..syft_object import SYFT_OBJECT_VERSION_1
-from ..syft_object import SyftObject
+from ...core.node.new.serializable import serializable
+from ...core.node.new.syft_object import SYFT_OBJECT_VERSION_1
+from ...core.node.new.syft_object import SyftObject
 
 
 @serializable(recursive_serde=True)
